Add relations (#134)
This PR implements relations for non-distributed workloads

* There are 2 different "explicit" relations: `Eval`, `Apply` generated
by actual `eval` and `apply`.
* Serializing a `Eval` or `Apply` relation serializes the lhs and rhs
objects of the relation and the relation itself, and any relation the
relation depends on.
* Serializing an object serializes the object itself (including any
recursively fully accessible objects), the relation itself and any
object the object "pins".
* The `rw-tester` serializes the `Eval` relation of the input Thunk.

This PR also implements a `fix` that interacts with `.fix` but does not
do any real work. `fix` currently has following commands:
* `what`:  print the handle to the object
* `content`: print content of an object
* `deep-content`: print content of an object recursively
* `result-apply` and `result-eval`: print the result of
eval/apply(object)
* `steps-eval`: print all dependencies of an eval-relation
* `step-eval`: print immediate dependencies of an eval-relation
* `explain`: print all relations point to the object + all tags that tag
the object + all objects this object pin

In addition, this PR adds a `short-name` arg to tester command line
args. The 7 bytes in the pretty-print of a Handle can be passed in and
expanded to the full handle