add fill as possible relation.
without fill as a possible relation, `get_source_relations` when trying
to `get_explanations` will throw.

a tree can be the target of a fill relation in the cache, so the
explanations must handle fill (either ignore or return it) as an
operation. this commit makes `get_explanations` return fill instead of
ignoring it.