Refactor Fix Ontology (#150)
Refactors the Fix type system to the new hierarchy, consisting of
Expressions, Objects, and Values, and adds Relations as a first-class
Fix type. Adds static checking for conversions within the Fix type
system. Separates components of Fixpoint out to make them easier to
test, and adds tests to validate correct behavior against the new
ontology. Implements an evaluator which applies the evaluation rules
under the new system, and exposes the new system to Fix programs via the
Wasm API surface.

Closes #138, #139, #140, #141, #142, #143, #144, #145, #146, and #147.

---------

Co-authored-by: Yuhan Deng <yhdeng@stanford.edu>