Add back functionalities of some old executables (#155)
This PR adds back some of the old executables' functionalities:
* `parse_args` matches new Handles, and `fix eval` perform similarly as
the old `stateless-tester` in terms of command line argument parsing.
After the execution is done, the generated `Eval` relation, and the two
sides of the relation are serialized to `Repository`, but other
relations that this `Eval` depends on is not yet serialized.
* Add `Client` and `Server`. Currently, `Client` only supports one
remote server and `Server` only supports one listening address. `Client`
puts `visit()` Handles to the remote server before sending a `get(
Handle<Eval> )` and `visit_minrepo()` Handles to the remote server
before sending a `get( Handle<Apply> )`.
* `NetworkWorker` and `Remote` buffer sending BlobData and TreeData
until sending a `Result` or `Run`. `ProposeTransfer and `AcceptTransfer
are sent before real data to eliminate duplicate sending.
* `Remote` now has a `view` of what the other side has. Any data
proposed by this side are considered existing on the other side after
the other side accepts the transfer. Any data proposed the other side
are considered existing after this side receives a proposal of transfer
of that data.
* Add back `fixpoint-client` and `fixpoint-server`.