Add WordCount Mapper and Reducer (#182)
This was a pain since declaring a `std::string` or `std::unordered_map`
(or any other non-POD C++ type) in the Wasm code would cause a linear
memory trap—I assume there's something more significant there going on
that needs to be debugged. This version should be valid in both C and
C++.

* the mapper uses quicksort to order the input words, then does a pass
over them to count and remove duplicates. It's probably around
`O(m*n*log(n))`, where `m` is the length of a word and `n` is the number
of words.
* the reducer effectively does mergesort to combine the lists, but also
sums/removes duplicates; it should be around `O(m*n)` per reduction