docs/fix.hs: proposed revision of Fix "model interpreter" semantics
At runtime, this would have one type of handle ("Expression"),
with three variants: Data, Thunk, and Encode.

Data can be a Ref or Object. Refs and Objects can be Blobs or Trees.

Thunks are opaque, but internally can be Application, Identification,
or Selection as before. When forced, they produce Data.

Encodes contain a Thunk, which gets forced to Data when the Encode is evaluated.
The Encode also contains an optional Bool that ensures a particular accessibility
of the Data.

`eval` is most similar to the current `reduce`, but less recursive:
it only produces a Value (a type with no *accessible* Encodes). It does
not lift the Refs.

Maybe the most controversial thing here is that `apply` and `select`
return an `Expression`, which means they can return an Encode. This means
that the force function explicitly ignores the accessibility field
of an "inner" Encode; the "outer" Encode's accessibility always wins.

The intention here is to let Encodes be used as tombstones if the storage
decides to physically discard a piece of Data. E.g. a Blob will be equivalent
to an Encode of a thunk that (when forced) produces that Blob. I *think*
this equivalency is maintained here but we'll want to prove that!